requirements.txt
torch
torchvision
pillow
flask
flask_sqlalchemy
werkzeug
opencv-python
numpy
ultralytics
tqdm
